+++ builder_pkgs/build_web_compilers/lib/src/dev_compiler_bootstrap.dart
@@ -60,20 +60,9 @@ Future<void> bootstrapDdc(
       throwIfUnsupported: !unsafeAllowUnsupportedModules,
     );
   } on UnsupportedModules catch (e) {
-    final librariesString = (await e.exactLibraries(buildStep).toList())
-        .map(
-          (lib) => AssetId(
-            lib.id.package,
-            lib.id.path.replaceFirst(moduleLibraryExtension, '.dart'),
-          ),
-        )
-        .join('\n');
-    log.warning('''
-Skipping compiling ${buildStep.inputId} with ddc because some of its
-transitive libraries have sdk dependencies that not supported on this platform:
-
-$librariesString
-''');
+    log.warning(
+      await e.formattedUnsupportedMessage('ddc', buildStep.inputId, buildStep),
+    );
     return;
   }
   final jsId = module.primarySource.changeExtension(jsModuleExtension);